requirements.txt
requests
openai
beautifulsoup4
pyyaml
